adding no action required option
