+++ definitions/benefit/sheets/AuthorisationCaseField/AuthorisationCaseField-nonprod.json
@@ -18,5 +18,6 @@
   {"LiveFrom": "25/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "tribunalCommunicationFilter", "AccessProfile": "caseworker-sscs-ibcaresponsewriter", "CRUD": "CRUD"},
   {"LiveFrom": "13/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "ftaRequestNoResponseRadioDl", "AccessProfile": "caseworker-sscs", "CRUD": "CRUD"},
   {"LiveFrom": "13/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "ftaRequestNoResponseQuery", "AccessProfile": "caseworker-sscs", "CRUD": "CRUD"},
-  {"LiveFrom": "13/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "ftaRequestNoResponseTextArea", "AccessProfile": "caseworker-sscs", "CRUD": "CRUD"}
+  {"LiveFrom": "13/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "ftaRequestNoResponseTextArea", "AccessProfile": "caseworker-sscs", "CRUD": "CRUD"},
+  {"LiveFrom": "13/03/2025", "CaseTypeID": "Benefit", "CaseFieldID": "ftaRequestNoResponseNoAction", "AccessProfile": "caseworker-sscs", "CRUD": "CRUD"}
 ]
